Add MemmoryMappedFileWordList.
Improve FileWordList by using BufferReader in #readFile.
Change the meaning of cachePercent to apply to file size rather than number of lines.
(It's meaning was never well defined.)
This allows the cache to be built inline while reading the file, removing the need to read the file twice.
This will generally result in larger caches, but users can tune the cache size down if that is an issue.