Add RuleResultMetadata component.
Metadata is used to communicate properties about the password which are related to the rule back to the validator.
This allows the validator to give feedback or advice to users attempting to create passwords which meet the rule requirements.
Note that metadata is returned for both valid and invalid password results.
Update WhitespaceRule to allow custom whitespace characters.