Relax PasswordGenerator API and add retry mech.
Relax generate API to accept a list of Rule so that rules that are naturally
part of password validation can also be specified for password generation.
The common use case is IllegalSequenceRule. Check the generated password
against the given ruleset and try again up to 2 times to generate a valid
password.

Resolves #152