v2: clear password data from memory
Prefer the use of CharSequence and CharBuffer in the API over String.
Write zeros to any buffers or arrays created by implementations.
Expand use of UnicodeString in the API in place of String.
Improve UnicodeString to support common string operations.
Add support for producing and testing heap dumps.
Add unit tests for UnicodeString and CharacterRule.
Fixes #126.