v2: Fail fast password validator.
Add failFast flag to DefaultPasswordValidator to exit validation when a rule fails.
This allows deployers to avoid expensive rule validation when it's not desired.
Move the CompositeRule from the test source to main to support more advanced grouping of rules.
Change the generics of entropy provider to support covariance of rules.
Fixes #172.